backend/requirements.txt
Flask
pdf2image
Pillow
python-dotenv
flask-cors